// src/app/models/Properties.ts
export interface Properties {
  zpid: number;
  id: number;
  providerListingId: string;
  imgSrc: string;
  hasImage: boolean;
  detailUrl: string;
  statusType: string;
  statusText: string;
  countryCurrency: string;
  price: string;
  unformattedPrice: number;
  address: string;
  addressStreet: string;
  addressCity: string;
  addressState: string;
  addressZipcode: number;
  isUndisclosedAddress: boolean;
  beds: number;
  baths: number;
  area: number;
  latitude: number;
  longitude: number;
  isZillowOwned: boolean;
  variableData: string;
  badgeInfo?: any;
  hdpData: string;
  isSaved: boolean;
  isUserClaimingOwner: boolean;
  isUserConfirmedClaim: boolean;
  pgapt: string;
  sgapt: string;
  zestimate?: number;
  shouldShowZestimateAsPrice: boolean;
  has3DModel: boolean;
  hasVideo: boolean;
  isHomeRec: boolean;
  info2String?: any;
  brokerName: string;
  hasAdditionalAttributions: boolean;
  isFeaturedListing: boolean;
  availabilityDate?: any;
  list: boolean;
  relaxed: boolean;
  info3String: string;
  hasOpenHouse?: boolean;
  openHouseStartDate?: Date;
  openHouseEndDate?: Date;
  openHouseDescription: string;
  builderName?: any;
  timeOfExtraction: number;
  extractionDate: string;
  homeType: string;
  daysOnZillow: number;
  isFeatured: boolean;
  datePriceChanged?: any;
  rentZestimate?: number;
  priceReduction?: any;
  isPreforclosureAuction: boolean;
  taxAssessedValue?: number;
  lotAreaValue?: number;
  lotAreaUnit: string;
  priceReductionStr: number;
  lotAreaRaw: number;
  priceReductionRatio: any;
  phone: any;
  Unnamed: number;
}